refactor: reviews 폴더의 컴포넌트 파일 구조 변경 #172
