+++ src/features/reviews/components/ReviewRating/SpoilerCheckBox.style.ts
